+++ resources/migrations/20211004130940-problems.up.sql
@@ -1,7 +1,7 @@
 create table problems (
   id serial primary key,
   num integer,
-  is_avail BOOLEAN default true,
+  is_avail integer default 1,
   problem text not null,
   test text default '',
   create_at timestamp default current_timestamp,